fix: issues with other extensions such as RLS (#19)
When changing operations we currently use the previous client, when this
happens any previous extensions are run again, which can cause issues.
For example prisma-extension-supabase-rls wraps the query in a
transaction and runs a raw sql query; when switching operations the query
hangs, maybe because the previous transaction's query never resolves or
becuase of some complication surrounding the tls setup query.

The query hooks are run after the previous extensions have been run, so
without an API to switch the operation type without rerunning previous
extension hooks it is not possible to fix this issue while continuing to
use them.

The only alternative I could find is to use model hooks instead,
replacing the query methods entirely and using the previous client for
the actual execution of the query. This prevents the issue of rerunning
the previous extensions because we only fire the query once we know
which operation we need.

Since we are no longer using query hooks the existing unit test
implementation no longer makes sense, since it depends on $allOperations
being run. Improve the realism of the MockClient so that we can write
the tests in a more e2e way: keep the previous client around and check
that it's methods are called correctly.