fix: fluent api
Closes #22

The changes done to fix RLS in #19 had an unfortunate side-effect of
breaking the Fluent API. Additionally the issues found in #19 seem to be
related to transactions, which are still not working correctly.

Revert the changes for #19 while still keeping the improvements to the
unit test setup.

BREAKING CHANGE: the previous client is no longer always used.

Users who use RLS extensions might encounter issues as the strategy of
always using the previous client is being reverted.